Reformatted the code and make it use for best ... Practices Created seperate files for each components and exported them
