+++ src/components/Header.jsx
@@ -1,0 +1,21 @@
+import headerImg from "../assets/react-core-concepts.png";
+
+const reactConcepts = ["Fundamental", "Crucial", "Core"];
+
+function getRandomInt(max) {
+  return Math.floor(Math.random() * (max + 1));
+}
+
+export default function Header() {
+  const keyConcepts = reactConcepts[getRandomInt(reactConcepts.length - 1)];
+  return (
+    <header>
+      <img src={headerImg} alt="Stylized atom" />
+      <h1>React Essentials</h1>
+      <p>
+        {keyConcepts} React concepts you will need for almost any app you are
+        going to build!
+      </p>
+    </header>
+  );
+}
